Swift 4 support for a 0.x release
[contains work from @ianpartridge and @weissi]

Motivation:

For an easier migration path, it is probably useful to ship a `0.x`
version of `swift-log` that is compatible with Swift 4. It can't use
any of the features that Swift 5 brings and therefore the Swift 4
version will have a slightly reduced feature set:

- log messages will be `String`
- logging metadata values will be `String` only (no dictionaries,
  arrays, ...)

That means, if you use the Swift 4 version of swift-log carefully
(only use string literals), your migration to the Swift 5 version should
only be a matter of changing your `Package.swift`.

Modifications:

- remove all string interpolation and use `String` directly

Result:

- we can release a 0.x version that supports Swift 4

Support Swift 4.1 (#40)

* Support Swift 4.1

* Set swift-tools-version to 4.0