Update metadata formatting for readability
Motivation:

For packages that generate logs and use them as part of the development process,
it is beneficial to be able to read through logs without having to write custom
log handlers or pull in a dependency for formatting only at dev time.

Modifications:

- Change: The format of `StreamLogHandler` to output metadata on newlines with tab indentation

Result:

Package authors will have an easier time debugging their applications without boilerplate code or dependencies.

This resolves #162